Move shared code from runtime to lib.
