+++ packages/graphql-entity/lib/graphql/loadDocumentsByConfig.ts
@@ -1,4 +1,4 @@
-import { resolveConfig } from './resolveConfig'
+import { resolveConfig } from '../config/resolveConfig'
 import { loadDocumentsFromGraphQLFiles } from './loadDocuments'
 
 export const loadDocumentsByConfig = (): ReturnType<typeof loadDocumentsFromGraphQLFiles> => {
